requirements.txt
Django
djangorestframework
psycopg2-binary
python-decouple
dj-database-url
gunicorn
whitenoise
